+++ server/utils/multer/index.js
@@ -1,0 +1,13 @@
+const multer = require("multer");
+const { avatarStorage } = require("./multerStorage");
+
+const uploadAvatar = multer({
+  storage: avatarStorage,
+  limits: {
+    fileSize: 1024 * 1024 * 5,
+  },
+});
+
+module.exports = {
+  uploadAvatar,
+};
